test(#3515): split the harness into tests/graceful_shutdown_support (file-size ratchet)
Lite FAILed the file-size component: 1664/1500 lines
(`cqlite-cli/tests/graceful_shutdown_tests.rs: 334 -> 1664`). Split per the
campsite rule (#1135), by responsibility rather than by line count:

* `tests/graceful_shutdown_support/mod.rs` (1262 lines) -- the INSTRUMENT: child
  I/O with a shared transcript, `WaitEnd`, calibrated `Budget`/`StageSpec`, the
  `StageClock`, the progress-checked poll, the bounded read-side SELECT, and the
  unit tests that pin the instrument's own invariants (floor invariant, cap-sum
  arithmetic, calibration baselines) -- they belong with the code they constrain.
* `tests/graceful_shutdown_tests.rs` (423 lines) -- the two integration tests
  that USE it, and nothing else.

A directory is not a test target, and the module has exactly one consumer, so the
parent's `#![cfg(all(feature = "write-support", unix))]` still gates it and no
other test target compiles it (which is why this did NOT go into the shared
`tests/common/`: every includer would compile it and trip dead_code under
`-D warnings`).

10/10 tests pass; `RUSTFLAGS="-D warnings" cargo clippy -p cqlite-cli --tests`
clean.
